Move ptxas/nvlink compilation into separate compilation unit
This moves all the PTX compilation functions that spawn subprocesses - notably ptxas, nvlink, and fatbin into a separate file.

The goal is to make this optional so that and eventually disable it by default. Since we can compile through libraries like libnvjitlink the rather brittle approach of calling external binaries is not needed anymore.

This also adds tests for all the helper functions. Tests for the actual compilation will follow separately.

PiperOrigin-RevId: 696102484